macipgw: Introduce a configuration file, GitHub #1549
- Read macipgw.conf from standard netatalk pkgconf path
- Introduce -f option to use custom config file path
- Avoid using u_long data type to appease musl

+++ contrib/macipgw/util.c
@@ -21,36 +21,37 @@
 #include <sys/types.h>
 
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 
 int gDebug;
 
 
-char *iptoa(u_long ip)
+char *iptoa(uint32_t ip)
 {
 	static char s[16];
 
-	sprintf(s, "%ld.%ld.%ld.%ld",
+	sprintf(s, "%u.%u.%u.%u",
 		(ip >> 24) & 0xff, (ip >> 16) & 0xff,
 		(ip >> 8) & 0xff, ip & 0xff);
 	return s;
 }
 
 
-u_long atoip(char *s)
+uint32_t atoip(char *s)
 {
-	u_long ip;
+	uint32_t ip;
 
-	ip = strtol(s, &s, 0);
+	ip = (uint32_t) strtol(s, &s, 0);
 	if (*s++ != '.')
 		return 0;
-	ip = (ip << 8) | strtol(s, &s, 0);
+	ip = (ip << 8) | (uint32_t) strtol(s, &s, 0);
 	if (*s++ != '.')
 		return 0;
-	ip = (ip << 8) | strtol(s, &s, 0);
+	ip = (ip << 8) | (uint32_t) strtol(s, &s, 0);
 	if (*s++ != '.')
 		return 0;
-	ip = (ip << 8) | strtol(s, &s, 0);
+	ip = (ip << 8) | (uint32_t) strtol(s, &s, 0);
 	if (*s != 0)
 		return 0;
 	return ip;
